requirements.txt
apache-beam==2.38.0
google-apitools
google-cloud-core==1.7.2
google-cloud-pubsub==2.8.0
google-cloud-storage==1.42.3
google-cloud-bigquery==2.34.2